qcacld-3.0: Update max value of cpu_cxpc_threshold
Current code has cpu_cxpc_threshold have a max value
of 17000, due to changes in newer targets, this value
has to be increased to accommodate.

CRs-Fixed: 3759288

+++ core/hdd/inc/wlan_hdd_cfg.h
@@ -204,7 +204,7 @@ struct hdd_config {
 	bool get_wifi_features;
 #endif
 #ifdef FEATURE_RUNTIME_PM
-	uint16_t cpu_cxpc_threshold;
+	unsigned long cpu_cxpc_threshold;
 #endif
 	bool exclude_selftx_from_cca_busy;
 #ifdef WLAN_FEATURE_11BE_MLO
